access restriction

+++ app/routes/post.py
@@ -1,4 +1,4 @@
-from flask import Blueprint, render_template, request, redirect
+from flask import Blueprint, render_template, request, redirect, abort
 from flask_login import login_required, current_user
 
 from ..models.post import Post
@@ -40,34 +40,41 @@ def create():
 @login_required
 def update(id):
     post = Post.query.get(id)
-    form = StudentForm()
-    form.student.data = User.query.filter_by(id=post.student).first().name
-    form.student.choices = [s.name for s in User.query.filter_by(status='user')]
-    if request.method == 'POST':
-        post.subject = request.form.get('subject')
-        student = request.form.get('student')
-        
-        post.student = User.query.filter_by(name=student).first().id
 
-        try:
-            db.session.commit()
-            return redirect('/')
-        except Exception as e:
-            print(str(e))
+    if post.author.id == current_user.id:
+        form = StudentForm()
+        form.student.data = User.query.filter_by(id=post.student).first().name
+        form.student.choices = [s.name for s in User.query.filter_by(status='user')]
+        if request.method == 'POST':
+            post.subject = request.form.get('subject')
+            student = request.form.get('student')
+            
+            post.student = User.query.filter_by(name=student).first().id
 
+            try:
+                db.session.commit()
+                return redirect('/')
+            except Exception as e:
+                print(str(e))
+
+        else:
+            return render_template('post/update.html', post=post, form=form)
     else:
-        return render_template('post/update.html', post=post, form=form)
+        abort(403)
 
 
 @post.route('/post/<int:id>/delete', methods=['POST', 'GET'])
 @login_required
 def delete(id):
     post = Post.query.get(id)
-        
-    try:
-        db.session.delete(post)
-        db.session.commit()
-        return redirect('/')
-    except Exception as e:
-        print(str(e))
-        return str(e)
+    
+    if post.author.id == current_user.id:
+        try:
+            db.session.delete(post)
+            db.session.commit()
+            return redirect('/')
+        except Exception as e:
+            print(str(e))
+            return str(e)
+    else:
+        abort(403)
